+++ src/e7.js
@@ -1,0 +1,18 @@
+import {curry} from 'katsu-curry'
+
+/**
+ * @method e7
+ * @param {string} fn - a function name
+ * @param {*} a - some parameter
+ * @param {*} b - some parameter
+ * @param {*} c - some parameter
+ * @param {*} d - some parameter
+ * @param {*} e - some parameter
+ * @param {*} f - some parameter
+ * @param {*} g - some parameter
+ * @param {Object} x - an object
+ * @returns {*}
+ */
+export const e7 = curry(
+  (fn, a, b, c, d, e, f, g, x) => x[fn](a, b, c, d, e, f, g)
+)
